# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "등대99",
    "description": "항해99 미니프로젝트, 등대99의 API를 정리해놓은 페이지"
  },
  "host": "localhost:3000", 
  "basePath": "/",
  "tags": [ ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in":"header",
      "description": "send JWT"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/study": {
      "get": {
        "tags": ["Study"],
        "description": "스터디 전체 목록 가져오기",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": ["Study"],
        "description": "스터디 업로드 하기",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "스터디 이름",
                  "example": "[급구] 스터디원 구합니다!"
                },
                "schedule": {
                  "type": "string",
                  "description": "스터디 일정",
                  "example": "약 4주간"
                },
                "startDate": {
                  "type": "string",
                  "description": "스터디 시작일",
                  "example": "2021-08-01"
                },
                "endJoinDate": {
                  "type": "string",
                  "description": "스터디 모집마감날짜",
                  "example": "2021-09-01"
                },
                "size": {
                  "type": "integer",
                  "description": "침여 가능 인원",
                  "example": 5
                },
                "explain": {
                  "type": "string",
                  "description": "스터디 상세내용",
                  "example": "실력 상관없습니다~"
                },
                "joinLater": {
                  "type": "boolean",
                  "description": "중도참여 여부",
                  "example": true
                },
                "userId": {
                  "type": "integer",
                  "description": "리더 인덱스",
                  "example": 1
                },
                "level": {
                  "type": "integer",
                  "description": "스터디 난이도 (상:0, 중:1, 하:2 무관:3)",
                  "example": 3
                },
                "studyType": {
                  "type": "integer",
                  "description": "스터디 방식 (감시형:0, 토이:1, 개인:2 독서실:3, 알고리즘:4)",
                  "example": 3
                },
                "joinNum": {
                  "type": "integer",
                  "description": "현재 참가 인원",
                  "example": 3
                },
                "writeDate": {
                  "type": "string",
                  "description": "스터디 업로드 일자",
                  "example": "2021-07-14"
                }
              }
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/study/{studyId}": {
      "get": {
        "tags": ["Study"],
        "description": "스터디 상세 내용 조회",
        "parameters": [
          {
            "name": "studyId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": ["Study"],
        "description": "스터디 수정",
        "parameters": [
          {
            "name": "studyId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "스터디 이름",
                  "example": "[일정 수정] 스터디원 구합니다!"
                },
                "schedule": {
                  "type": "string",
                  "description": "스터디 일정",
                  "example": "약 2주간"
                },
                "startDate": {
                  "type": "string",
                  "description": "스터디 시작일",
                  "example": "2021-08-15"
                },
                "endJoinDate": {
                  "type": "string",
                  "description": "스터디 모집마감날짜",
                  "example": "2021-09-01"
                },
                "writeDate": {
                  "type": "string",
                  "description": "스터디 업로드 일자",
                  "example": "2021-07-14"
                },
                "size": {
                  "type": "integer",
                  "description": "침여 가능 인원",
                  "example": 5
                },
                "explain": {
                  "type": "string",
                  "description": "스터디 상세내용",
                  "example": "일정 변경 되었습니다!"
                },
                "joinLater": {
                  "type": "boolean",
                  "description": "중도참여 여부",
                  "example": true
                },
                "level": {
                  "type": "integer",
                  "description": "스터디 난이도 (상:0, 중:1, 하:2 무관:3)",
                  "example": 3
                },
                "studyType": {
                  "type": "integer",
                  "description": "스터디 방식 (감시형:0, 토이:1, 개인:2 독서실:3, 알고리즘:4)",
                  "example": 3
                },
                "joinNum": {
                  "type": "integer",
                  "description": "현재 참가 인원",
                  "example": 3
                }
              }
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": ["Study"],
        "description": "스터디 삭제",
        "parameters": [
          {
            "name": "studyId",
            "in": "path",
            "description": "스터디 인덱스",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/join-study/{studyId}": {
      "post": {
        "tags": ["Study Join"],
        "description": "스터디 신청하기",
        "parameters": [
          {
            "name": "studyId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "integer",
                  "description": "참가자의 user 인덱스",
                  "example": 1
                },
                "leader": {
                  "type": "integer",
                  "description": "스터디 리더의 user 인덱스",
                  "example": 2
                }
              }
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "tags": ["Study Join"],
        "description": "스터디 신청 내역 인덱스",
        "parameters": [
          {
            "name": "studyId",
            "in": "path",
            "description": "스터디 인덱스",
            "required": true,
            "type": "integer"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "interger",
                  "description": "참여자의 user 인덱스",
                  "example": 1
                }
              }
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/study-all-comment/{studyId}": {
      "get": {
        "tags": ["Study-Comment"],
        "description": "각 스터디 별 댓글",
        "parameters": [
          {
            "name": "studyId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    
    "/api/study-comment": {
      "post": {
        "tags": ["Study-Comment"],
        "description": "각 스터디 별 댓글",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "integer",
                  "description": "댓글 작성자의 user 인덱스",
                  "example": 1
                },
                "studyId": {
                  "type": "integer",
                  "description": "스터디 인덱스",
                  "example": 2
                },
                "content": {
                  "type": "string",
                  "description": "댓글 내용",
                  "example": "스터디 내용이 너무 알차요!"
                }
              }
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },

    "/api/study-comment/{studyCommentId}": {
      "delete": {
        "tags": ["Study-Comment"],
        "description": "댓글 삭제하기",
        "parameters": [
          {
            "name": "studyCommentId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["Study-Comment"],
        "description": "댓글 수정하기",
        "parameters": [
          {
            "name": "studyCommentId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "description": "댓글 수정 내용",
                  "example": "알차네요!"
                }
              }
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/recent-study": {
      "get": {
        "tags": ["Main-Page"],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/register": {
      "post": {
        "tags": ["User"],
        "description": "회원가입",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "description": "이메일",
                  "example": "test@email.com"
                },
                "nickname": {
                  "type": "string",
                  "description": "닉네임",
                  "example": "테스트계정"
                },
                "group": {
                  "type": "integer",
                  "description": "항해 기수",
                  "example": 2
                },
                "password": {
                  "type": "string",
                  "description": "비밀번호",
                  "example": "password123"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/login": {
      "post": {
        "tags": ["User"],
        "description": "로그인",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "description": "이메일",
                  "example": "test@email.com"
                },
                "password": {
                  "type": "string",
                  "description": "비밀번호",
                  "example": "password123"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/briefInfo/{userId}": {
      "get": {
        "tags": ["My-Info"],
        "description": "마이페이지",
        "security": {
          "bearerAuth": []
        },
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/mystudy/{userId}": {
      "get": {
        "tags": ["My-Info"],
        "description": "내 스터디 목록 조회하기",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/myinfo/{userId}": {
      "put": {
        "tags": ["My-Info"],
        "description": "내 정보 조회",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nickname": {
                  "type": "string",
                  "description": "닉네임",
                  "example": "테스트계정"
                },
                "password": {
                  "type": "string",
                  "description": "비밀번호",
                  "example": "password123"
                }
              }
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/mypost/{userId}": {
      "get": {
        "tags": ["My-Info"],
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/mycomment/{userId}": {
      "get": {
        "tags": ["My-Info"],
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "JWT 토큰 값 (앞에 Bearer 포함 필수)",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "definitions": {
    "Study": {
      "type": "object",
      "properties": {
        "studyId":{
          "type":"Number"
        },
        "name":{
          "type":"String"
        },
        "schedule":{
          "type":"String"
        },
        "startDate":{
          "type":"String"
        },
        "endJoinDate":{
          "type":"String"
        },
        "writeDate":{
          "type":"String"
        },
        "size":{
          "type":"Number"
        },
        "explain":{
          "type":"String"
        },
        "joinLater":{
          "type":"Boolean"
        },
        "userId":{
          "type":"Number"
        },
        "level":{
          "type":"Number"
        },
        "studyType":{
          "type":"Number"
        },
        "joinNum":{
          "type":"Number"
        }
    }
  },
    "StudyComment": {
      "type": "object",
      "properties": {
        "studyCommentId":{
          "type":"Number"
        },
        "studyId":{
          "type":"Number"
        },
        "content":{
          "type":"String"
        },
        "userId":{
          "type":"String"
        },
        "date":{
          "type":"String"
        }
      }
    },
    "StudyJoin": {
      "type": "object",
      "properties": {
        "studyId":{
          "type":"Number"
        },
        "userId":{
          "type":"Number"
        },
        "userName":{
          "type":"String"
        },
        "leader":{
          "type":"Boolean"
        }
      }
    }, 
     "User": {
      "type": "object",
      "properties": {
        "userId":{
        "type":"Number"
      },
      "group":{
        "type":"Number"
      },
      "email":{
        "type":"String"
      },
      "nickname":{
        "type":"String"
      },
      "password":{
        "type":"String"
      }
    }
    }
  }
}